n1 kernel bump 5.15 to 5.15.135
This commit updates the LINUX_VERSION and LINUX_KERNEL_HASH variables in the diy-part1.sh script. The Kernel version is updated from .133 to .135 and its corresponding hash is also updated. These changes are essential for ensuring the script aligns with the updated Kernel version and maintain the system's optimal function.